Changed models from a file to a package.
