+++ statler_api/models/__init__.py
@@ -1,0 +1,7 @@
+"""Models which represent Statler's objects from the database to JSON and back."""
+
+# Explicitly importing package contents here makes them available as models.Play rather than models.play.Play
+from .play import PlayDAO
+from .playlist import PlayListDAO
+from .playlistentry import PlayListEntryDAO
+from .review import ReviewDAO
